test_airflow/requirements.txt
pandas 
pymongo
emoji
pandas
psycopg2-binary
python-dotenv
requests
tabulate
tqdm
PyYAML
redis